+++ .gitattributes
@@ -1,5 +1,6 @@
 # ---> Vendored Files according to Linguist https://stackoverflow.com/a/26948938/7061105
 clean-up-repository-by-gitignore.sh linguist-vendored
 generate-README.sh linguist-vendored
+git_init-branch.sh linguist-vendored
 .asciidoc/* linguist-vendored
 docs/* linguist-vendored
