+++ app/src/main/kotlin/ooo/akito/webmon/data/db/DatabaseTypeConvertorList.kt
@@ -1,0 +1,3 @@
+package ooo.akito.webmon.data.db
+
+class DatabaseTypeConvertorList : DatabaseTypeConvertorBase<String>()
